add a few srcs, cmplete Board::readWeight, readDef

+++ src/Coor.cpp
@@ -4,31 +4,31 @@ Coor::Coor() :
   x(-1), 
   y(-1){}
 
-Coor::Coor(double X, double Y){
+Coor::Coor(float X, float Y){
   x = X;
   y = Y;
 }
 
 Coor::~Coor(){}
 
-void Coor::setX(double X){
+void Coor::setX(float X){
   x = X;
 }
 
-void Coor::setY(double Y){
+void Coor::setY(float Y){
   y = Y;
 }
 
-void Coor::setCoor(double X, double Y){
+void Coor::setCoor(float X, float Y){
   x = X;
   y = Y;
 }
 
-double Coor::getX() const{
+float Coor::getX() const{
   return x;
 }
 
-double Coor::getY() const{
+float Coor::getY() const{
   return y;
 }
